+++ Globals.py
@@ -4,10 +4,11 @@ _VERSION = 'alpha 0.3.6'
 _TITLE = f"ToN Codes {_VERSION} - by MichelleDB"
 _WIDTH = 620
 _HEIGHT = 300
+_GITHUB = 'https://github.com/69MichelleDB/ToNCodes'
 
 # Options window
 _WIDTH_OPT = 550
-_HEIGHT_OPT = 140
+_HEIGHT_OPT = 180
 
 # About window
 _WIDTH_ABOUT = 500
@@ -21,6 +22,7 @@ _FOLDER_TEMP = 'Temp'
 # Global variables
 configList = {}
 root = None
+titleMessage = ''
 
 # TEMP
 auxPathFirstBoot = None