+++ ToNCodes.py
@@ -1,7 +1,8 @@
-from Tools.xmlTools import InitializeConfig, ReadCodeFiles
+from Tools.xmlTools import InitializeConfig, ReadCodeFiles, ModifyNode
 from Tools.fileTools import VerifyInitFileStructure, CreateFromDefault, CleanTempFiles
 from Tools.guiTools import CreateWindow, HorizontalMenu, CreateTreeView, CalculatePosition
 from Tools.errorHandler import ErrorLogging
+from Tools.webhookTool import CheckForUpdates
 from CodesHunter import CodesHunter
 import threading
 import Globals as gs
@@ -25,27 +26,26 @@ def RefreshCodes():
 
 
 if __name__ == "__main__":
-    try:
-        VerifyInitFileStructure()
+    VerifyInitFileStructure()
 
-        CleanTempFiles()
+    CleanTempFiles()
 
-        CreateFromDefault(gs._CONFIG_FILE)                             # Verify the config file exist
-        gs.configList = InitializeConfig(gs._CONFIG_FILE)              # Retrieve all config data
+    CreateFromDefault(gs._CONFIG_FILE)                              # Verify the config file exist
+    gs.configList = InitializeConfig(gs._CONFIG_FILE)               # Retrieve all config data
 
-        # The main program to check for codes and create the XML files
-        mainThread = threading.Thread(target=CodesHunter)
-        mainThread.daemon = True
-        mainThread.start()
+    CheckForUpdates()                                               # Check for new updates logic
 
-        # GUI setup
-        gs.root = CreateWindow(gs._TITLE, gs._WIDTH, gs._HEIGHT, True)
-        auxX,auxY = CalculatePosition(gs._WIDTH, gs._HEIGHT)
-        gs.root.geometry(f'{gs._WIDTH}x{gs._HEIGHT}+{auxX}+{auxY}')
-        HorizontalMenu(gs.root)
-        codesData = RefreshCodes()
-        CreateTreeView(gs.root, codesData, int(gs.configList['gui-delay'])*1000, RefreshCodes)
+    # The main program to check for codes and create the XML files
+    mainThread = threading.Thread(target=CodesHunter)
+    mainThread.daemon = True
+    mainThread.start()
 
-        gs.root.mainloop()
-    except Exception as e:
-        ErrorLogging(f"Error in Main: {e}")
+    # GUI setup
+    gs.root = CreateWindow(gs._TITLE + gs.titleMessage, gs._WIDTH, gs._HEIGHT, True)
+    auxX,auxY = CalculatePosition(gs._WIDTH, gs._HEIGHT)
+    gs.root.geometry(f'{gs._WIDTH}x{gs._HEIGHT}+{auxX}+{auxY}')
+    HorizontalMenu(gs.root)
+    codesData = RefreshCodes()
+    CreateTreeView(gs.root, codesData, int(gs.configList['gui-delay'])*1000, RefreshCodes)
+
+    gs.root.mainloop()